requirements.txt
numpy
opencv-python
Pillow
torch
torchvision>=0.8.0
tqdm
huggingface-hub
tensorflow
torchaudio
rembg
flask
functools
opencv-python
os
werkzeug
sqlite3
wandb
glob
matplotlib
zipfile